Separate out SiP code
